# backend/cmd/docs/swagger.yaml
basePath: /
definitions:
  entity.Customer:
    properties:
      id:
        type: integer
      is_public:
        type: integer
      name:
        type: string
      sent_warning:
        type: string
      wage:
        type: number
    type: object
  entity.User:
    properties:
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      pass:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@email.com
    name: API Support
    url: http://www.contact.url
  description: Projeto final para o AceleraDev FullStack presencial CodeNation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://terms.of.service
  title: Captação de Clientes Banco Uati
  version: "0.1"
paths:
  /customer:
    post:
      consumes:
      - application/json
      description: Adiciona um cliente ao BD
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Not'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Bad'
            type: object
      summary: Adiciona um cliente ao BD
      tags:
      - customer
  /user:
    post:
      consumes:
      - application/json
      description: Adiciona um usuário ao BD
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/URL'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Erro'
            type: object
      summary: Adiciona um usuário ao BD
      tags:
      - user
securityDefinitions:
  'ApiKeyAuth TODO: not implemented yet':
    in: 'header TODO: not implemented yet'
    name: 'Authorization TODO: not implemented yet'
    type: apiKey
  'OAuth2AccessCode TODO: not implemented yet':
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  'OAuth2Application TODO: not implemented yet':
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  'OAuth2Implicit TODO: not implemented yet':
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  'OAuth2Password TODO: not implemented yet':
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  'TODO: not implemented yet':
    type: basic
swagger: "2.0"